requirements/work.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
gunicorn==21.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssselect==1.2.0
defusedxml==0.7.1
Django==4.2.8
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
django-solo==2.1.0
django-tinymce==3.6.1
djangorestframework==3.14.0
drf-spectacular==0.26.1
fonttools==4.46.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
lxml==5.1.0
packaging==23.2
parsel==1.8.1
phonenumbers==8.13.27
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.12
